# core/grantforge-web/src/api/openapi.json
{
  "components" : {
    "schemas" : {
...
      "LoginHistoryResponse" : {
        "properties" : {
          "action" : {
            "enum" : [ "LOGIN_SUCCEEDED", "LOGIN_FAILED", "ACCOUNT_LOCKED", "LOGOUT", "SESSION_REVOKED", "PASSWORD_CHANGED", "TENANT_CREATED", "TENANT_UPDATED", "TENANT_SUSPENDED", "TENANT_ACTIVATED" ],
            "type" : "string"
          },
          "clientIp" : {
            "type" : "string"
          },
          "occurredAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "outcome" : {
            "enum" : [ "SUCCESS", "FAILURE" ],
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          },
          "userAgent" : {
            "type" : "string"
          }
        },
        "required" : [ "action", "occurredAt", "outcome" ],
        "type" : "object"
      },
...
      "MeResponse" : {
        "properties" : {
          "displayName" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "lastLoginAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "passwordChangeRequired" : {
            "type" : "boolean"
          },
          "systemAccount" : {
            "type" : "boolean"
          },
          "tenantCode" : {
            "type" : "string"
          },
          "tenantName" : {
            "type" : "string"
          },
          "username" : {
            "type" : "string"
          }
        },
        "required" : [ "passwordChangeRequired", "systemAccount", "tenantCode", "tenantName", "username" ],
        "type" : "object"
      },
      "PageResultLoginHistoryResponse" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/LoginHistoryResponse"
            },
            "type" : "array"
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "total" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "items", "page", "size", "total" ],
        "type" : "object"
      },
      "PageResultSessionResponse" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/SessionResponse"
            },
            "type" : "array"
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "total" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "items", "page", "size", "total" ],
        "type" : "object"
      },
...
      "PasswordChangeRequest" : {
        "properties" : {
          "currentPassword" : {
            "maxLength" : 1024,
            "minLength" : 0,
            "type" : "string"
          },
          "newPassword" : {
            "maxLength" : 1024,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "currentPassword", "newPassword" ],
        "type" : "object"
      },
      "ProfileRequest" : {
        "properties" : {
          "displayName" : {
            "maxLength" : 128,
            "minLength" : 0,
            "type" : "string"
          },
          "email" : {
            "maxLength" : 254,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "SessionResponse" : {
        "properties" : {
          "clientIp" : {
            "type" : "string"
          },
          "current" : {
            "type" : "boolean"
          },
          "displayName" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "lastSeenAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "signedInAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "userAgent" : {
            "type" : "string"
          },
          "username" : {
            "type" : "string"
          }
        },
        "required" : [ "current", "id", "lastSeenAt", "signedInAt", "username" ],
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "description" : "Fine-grained, plugin-based authorization platform",
    "license" : {
      "name" : "MIT",
      "url" : "https://opensource.org/license/mit"
    },
    "title" : "GrantForge API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/me" : {
      "get" : {
        "operationId" : "me",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MeResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "me-controller" ]
      },
      "put" : {
        "operationId" : "update_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProfileRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MeResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "me-controller" ]
      }
    },
...
    "/api/v1/me/login-history" : {
      "get" : {
        "operationId" : "loginHistory",
        "parameters" : [ {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResultLoginHistoryResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "me-controller" ]
      }
    },
    "/api/v1/me/password" : {
      "post" : {
        "operationId" : "changePassword",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PasswordChangeRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "me-controller" ]
      }
    },
...
    "/api/v1/me/sessions/{id}" : {
      "delete" : {
        "operationId" : "revokeOwn",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "session-controller" ]
      }
    },
    "/api/v1/sessions" : {
      "get" : {
        "operationId" : "all",
        "parameters" : [ {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResultSessionResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "session-controller" ]
      }
    },
    "/api/v1/sessions/{id}" : {
      "delete" : {
        "operationId" : "revoke",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "session-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "url" : "/"
  } ]
}